requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
bokeh==3.6.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorcet==3.1.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cucim-cu12==25.4.0
cuda-bindings==12.8.0
cuda-python==12.8.0
cudf-cu12==25.4.0
cugraph-cu12==25.4.1
cuml-cu12==25.4.0
cuproj-cu12==25.4.0
cupy-cuda12x==13.4.1
cuspatial-cu12==25.4.0
cuvs-cu12==25.4.0
cuxfilter-cu12==25.4.0
cycler==0.12.1
dask==2025.2.0
dask-cuda==25.4.0
dask-cudf-cu12==25.4.0
datashader==0.18.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distributed==2025.2.0
distributed-ucxx-cu12==0.43.0
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
geopandas==1.0.1
greenlet==3.2.1
h11==0.16.0
holoviews==1.20.2
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
libcudf-cu12==25.4.0
libcugraph-cu12==25.4.1
libcuml-cu12==25.4.0
libcuspatial-cu12==25.4.0
libcuvs-cu12==25.4.0
libkvikio-cu12==25.4.0
libraft-cu12==25.4.0
librmm-cu12==25.4.0
libucx-cu12==1.18.1
libucxx-cu12==0.43.0
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multipledispatch==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.1
notebook_shim==0.2.4
numba==0.60.0
numba-cuda==0.4.0
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.26.2
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.11
nx-cugraph-cu12==25.4.0
opencv-python==4.11.0.86
optuna==4.3.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
panel==1.6.3
param==2.2.0
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pyct==0.5.0
Pygments==2.19.1
pylibcudf-cu12==25.4.0
pylibcugraph-cu12==25.4.1
pylibraft-cu12==25.4.0
pynvjitlink-cu12==0.6.0
pynvml==12.0.0
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.4.0
raft-dask-cu12==25.4.0
rapids-dask-dependency==25.4.0
rapids-logger==0.1.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rmm-cu12==25.4.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.3.1
shapely==2.0.7
simpervisor==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.14.0
tblib==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
toolz==1.0.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelite==4.4.1
triton==3.3.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
ucx-py-cu12==0.43.0
ucxx-cu12==0.43.0
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xarray==2025.4.0
xyzservices==2025.4.0
yarl==1.20.0
zict==3.0.0
